Implement host keys message & proving for server side
This implements the hostkeys-00@openssh.com message sent by the server
and the subsequent reply for the hostkeys-prove-00@openssh.com message
from the client.

The SSH_BIND_OPTIONS_HOSTKEY usage is slight extended here. The behavior
of last configured key wins still applies, but any previously applied
key is kept for additional host key verification.

This extension allows for implementation a key rotation scheme in a
libssh server, by adding the new key first and then the current key
(even for the same algorithm like RSA), so that both the new and current
key are sent during host prove messages.

Once the new key has been added like this for a while and clients have
updated the host keys, the latter call of SSH_BIND_OPTIONS_HOSTKEY can
be removed to stop using the old key and fall back to the new one that
clients now know about.

OpenSSH has enabled UpdateHostkeys by default for the next version, also
see
https://github.com/openssh/openssh-portable/commit/1286981d08b8429a64613215ce8bff3f6b32488a,
which is going to be part of the next OpenSSH release.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>