test: Refactor test server callback for authentication
This refactors the default callback to be better suited for future host
keys rotation. The host keys need to be sent from the server side after
authentication completes.

Before the authenticated state was already set even with the NONE public
key handshake, which is not authentication but only a check if public
key authentication is allowed.

This changes the callback to only set the authenticated state once it
really completes. It also splits the authentication part from the
channel setup, so that host keys can be added in follow up commits.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>