Add client side support for hostkeys rotation
This implements client side support for hostkeys rotation. It ends up
tracking the final verified host keys in the session, and the logic for
updating host keys handles these as well.

In general, the host keys update logic is changed so it doesn't add any
host keys that are already known so that it is safe to call this method
after authenticated and handshakes complete, so that host keys can be
updated / added if they are received from the server.

It adds a new function to the API that allows for replying to received
host keys to ask for proof. This function can optionally wait for a bit
of no message have been received. This happens if the method is called
right after then authentication completes and the host keys are still in
flight from the server to the client.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>