test: Add server integration tests for host keys
This adds an end to end integration test for server side handling of
host keys. This depends primarily on client support now and we verify on
the client side that keys are received and verified properly.

Verifying on the client implies that the server both receives the keys
and is able to sign them with a correct proof and signature.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>