Add ability to retrieve last PKI related error message
Add a API function ssh_pki_get_error to retrieve the last
PKI related error message. Many PKI related error messages
are recoverable errors that ought to be displayed to a user,
e.g. the wrong passphrase for a key.

Other solutions like setting the overall log to SSH_LOG_TRACE
and displaying all logs to a user (or making a user read a log
file, or parsing the log messages in a callback to determine what
to display, etc.) are less than ideal.

Note that setting an error sends the same message to the log at
SSH_LOG_TRACE level.

Fix #232

Signed-off-by: John Thacker <johnthacker@gmail.com>