example: Fix insufficient validation of the head of cleanup node list
In the previous commit(8164e1ff9cba), the NULL check for head_pre is
removed and the NULL check for *head_pre is added. But the new check
doesn't really work, because *head_pre can be assigned to new_node->next,
whether it's NULL or not. In addition, *head_pre is not checked before
the dereference below.

The cleanup_push() is used to add new_data to a cleanup node list.
The *head_pre being NULL means that the list does not exist, which
makes cleanup_push meaningless. Thus, the *head_pre cannot be NULL, when
cleanup_push() is called.

Signed-off-by: renmingshuai <renmingshuai@huawei.com>